New tests/ and toolkit/ folder. Moving files
